Se crea la vista para el cuadro en PDF y se modifica la base de datos para que almacene las categorias superior e inferior

+++ database/migrations/2019_02_07_165902_solicitudes_solicitud.php
@@ -22,7 +22,7 @@ class SolicitudesSolicitud extends Migration
             $table->string('SOLICITUD_PUESTO');
             $table->text('SOLICITUD_ACTIVIDADES');
             $table->string('SOLICITUD_NOMINA');
-            $table->integer('SOLICITUD_SALARIO');
+            $table->float('SOLICITUD_SALARIO');
             $table->text('SOLICITUD_JUSTIFICACION');
             $table->string('SOLICITUD_TIPO_SOLICITUD');
             $table->timestamps();
